Stop executing flaky tests that can fail because of aggressive GC (#28660)
